test_api/requirements.txt
# Extra deps for test_api scripts only
google-generativeai>=0.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
